set maximum possible zoomlevel to 28 optimized Textbuffer layout increased circle/rect/img coordinate precision

+++ packages/display/src/displays/webgl/buffer/templates/TextBuffer.ts
@@ -27,7 +27,7 @@ export class TextBuffer extends TemplateBuffer {
         this.attributes = {
             // vertex
             a_position: {
-                data: new FlexArray(Float32Array),
+                data: new FlexArray(Int16Array),
                 size: 2
             },
             // point
